requirements.txt
asn1crypto==0.24.0
astroid==1.6.6
awscli==1.17.5
backports.functools-lru-cache==1.6.1
botocore==1.14.5
CodeIntel==0.9.3
colorama==0.4.1
configparser==4.0.2
cryptography==2.1.4
Django==1.11.10
docutils==0.15.2
enum34==1.1.6
futures==3.3.0
idna==2.6
ikpdb==1.2.3
Inflector==2.0.12
ipaddress==1.0.17
isort==4.3.21
jedi==0.11.1
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.1.1
pbr==5.4.4
pyasn1==0.4.8
pycrypto==2.6.1
pygobject==3.26.1
pylint==1.9.5
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
pyxdg==0.25
PyYAML==5.2
rsa==3.4.2
s3transfer==0.3.1
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.14.0
stevedore==1.31.0
subprocess32==3.5.4
urllib3==1.25.7
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
zope.cachedescriptors==4.3.1
